More files

+++ src/Controller/WalletController.php
@@ -6,12 +6,15 @@ use App\Entity\CronTask;
 use App\Entity\Wallet;
 use App\Form\WalletType;
 use App\Repository\WalletRepository;
-use App\Service\Kraken;
+use App\Service\BChain;
+use App\Service\BlockChain;
+use App\Service\CronTab;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
+use Symfony\Component\Security\Core\Security;
 
 /**
  * @Route("/wallet")
@@ -19,10 +22,16 @@ use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 class WalletController extends AbstractController
 {
     private $authChecker;
+    private $blockchain;
+    private $cronTab;
+    private $security;
 
-    public function __construct(AuthorizationCheckerInterface $authChecker)
+    public function __construct(CronTab $cronTab, BChain $blockChain, AuthorizationCheckerInterface $authChecker, Security $security)
     {
         $this->authChecker = $authChecker;
+        $this->blockchain = $blockChain;
+        $this->cronTab = $cronTab;
+        $this->security = $security;
     }
 
     /**
@@ -52,6 +61,7 @@ class WalletController extends AbstractController
             if ($form->isSubmitted() && $form->isValid()) {
 
                 $entityManager = $this->getDoctrine()->getManager();
+                $bchain = null;
 
                 /**
                  * 1. connexion à la blockchain
@@ -66,21 +76,47 @@ class WalletController extends AbstractController
                  *
                  */
 
-                $bchain = $this->container->get('blockchain')->getChain($wallet);
+                try
+                {
+                    $bchain = $this->blockchain->getChain($wallet);
+                }catch (\Exception $e)
+                {
+                    // Manage error
+                    return $this->render('wallet/new.html.twig', [
+                        'wallet' => $wallet,
+                        'form' => $form->createView(),
+                        'error' => "Unable to connect to the API -- ".$e->getMessage()
+                    ]);
+                }
 
-                //Création du CronTask lié à ce wallet
-                $cron = new CronTask();
-                $cron->setEtat(0);//Inactif de base
-                $cron->setWallet($wallet);
+                if($bchain)
+                {
+                    //Création du CronTask lié à ce wallet
+                    $cron = new CronTask();
+                    $cron->setWallet($wallet);
 
-                $entityManager->persist($wallet);
-                $entityManager->persist($cron);
+                    $entityManager->persist($wallet);
+                    $entityManager->persist($cron);
 
-                var_dump($form);
+                    //Activate Crontask
+                    //Activate trading cron
+                    $this->cronTab->create($wallet);
+                    
+                    $wallet->setUser($this->security->getUser());
 
-                /*$entityManager->flush();
+                    $entityManager->flush();
 
-                return $this->redirectToRoute('wallet_index');*/
+                    return $this->redirectToRoute('wallet_show',['id'=>$wallet->getId()]);
+
+                }else
+                {
+                    // Manage error
+                    return $this->render('wallet/new.html.twig', [
+                        'wallet' => $wallet,
+                        'form' => $form->createView(),
+                        'error'=>'blochain not created'
+                    ]);
+                }
             }
 
             return $this->render('wallet/new.html.twig', [
@@ -93,21 +129,79 @@ class WalletController extends AbstractController
     }
 
     /**
+     * @Route("/active_trading", name="active_trading", methods={"POST"})
+     */
+    public function activateAction(Request $request)
+    {
+        $data = array();
+
+        if($request->isMethod('POST'))
+        {
+            $done = array();
+            $em = $this->getDoctrine()->getManager();
+
+            foreach (explode(",",$request->request->get('on')) as $signature)
+            {
+                $wallet = $em->getRepository('App:Wallet')->findOneBy(['signature'=>$signature]);
+                if($wallet && $wallet->getTrading() != 1)
+                {
+                    $wallet->setTrading(1);
+                    $done[]=$wallet->getName();
+
+                    //Activate trading cron
+                    $this->cronTab->create($wallet);
+                }
+            }
+
+            foreach (explode(",",$request->request->get('off')) as $signature)
+            {
+                $wallet = $em->getRepository('App:Wallet')->findOneBy(['signature'=>$signature]);
+                if($wallet && $wallet->getTrading() != 0)
+                {
+                    $wallet->setTrading(0);
+                    $done[]=$wallet->getName();
+
+                    //Desactivate trading cron
+                    $this->cronTab->remove($wallet);
+                }
+            }
+
+            $em->flush();
+
+            $data = [
+                'result'=>1,
+                'updated'=>implode(",", $done)
+            ];
+
+        }
+
+        return $this->json($data);
+    }
+
+    /**
      * @Route("/{id}", name="wallet_show", methods={"GET"})
      */
     public function show(Wallet $wallet): Response
     {
-        if ($this->authChecker->isGranted('IS_AUTHENTICATED_REMEMBERED')) {
+        if ($this->authChecker->isGranted('IS_AUTHENTICATED_REMEMBERED') &&
+            $this->security->getUser()->getId() == $wallet->getUser()->getId() ) {
+            $trade = $this->blockchain->getChain($wallet);
+            if($trade) {
 
-            $trade = $this->container->get('blockchain')->getChain($wallet);
-            $history = $trade->getHistory('X' . $wallet->getCryptomonnaie()->getSymbol(), 'EUR', new \DateTime('03/09/2015'));
+                $history = $trade->getHistory('X' . $wallet->getCurrency()->getSymbol(), 'EUR', new \DateTime('03/09/2015'));
 
-            return $this->render('wallet/show.html.twig', [
-                'wallet' => $wallet,
-                'history' => $history
-            ]);
+                return $this->render('wallet/show.html.twig', [
+                    'wallet' => $wallet,
+                    'history' => $history
+                ]);
+            }else{
+                return $this->render('wallet/show.html.twig', [
+                    'wallet' => $wallet,
+                    'history' => array()
+                ]);
+            }
         }else
-            return $this->render('user_login');
+            return $this->redirectToRoute('user_login');
     }
 
     /**
@@ -115,7 +209,9 @@ class WalletController extends AbstractController
      */
     public function edit(Request $request, Wallet $wallet): Response
     {
-        if ($this->authChecker->isGranted('IS_AUTHENTICATED_REMEMBERED')) {
+        if ( $this->authChecker->isGranted('IS_AUTHENTICATED_REMEMBERED') &&
+             $this->security->getUser()->getId() == $wallet->getUser()->getId() )
+        {
             $form = $this->createForm(WalletType::class, $wallet);
             $form->handleRequest($request);
             $cron = null;
@@ -131,14 +227,13 @@ class WalletController extends AbstractController
                 if (!$cron) {
                     //Création du CronTask lié à ce wallet
                     $cron = new CronTask();
-                    $cron->setEtat(0);//Inactif de base
                     $cron->setWallet($wallet);
                     $em->persist($cron);
                 }
 
                 $em->flush();
 
-                return $this->redirectToRoute('wallet_index');
+                return $this->redirectToRoute('wallet_show',['id'=>$wallet->getId()]);
             }
 
             return $this->render('wallet/edit.html.twig', [
@@ -146,7 +241,7 @@ class WalletController extends AbstractController
                 'form' => $form->createView(),
             ]);
         }else
-            return $this->render('user_login');
+            return $this->redirectToRoute('user_login');
     }
 
     /**
@@ -154,7 +249,8 @@ class WalletController extends AbstractController
      */
     public function delete(Request $request, Wallet $wallet): Response
     {
-        if ($this->authChecker->isGranted('IS_AUTHENTICATED_REMEMBERED')){
+        if ($this->authChecker->isGranted('IS_AUTHENTICATED_REMEMBERED') &&
+            $this->security->getUser()->getId() == $wallet->getUser()->getId() ){
             if ($this->isCsrfTokenValid('delete' . $wallet->getId(), $request->request->get('_token'))) {
                 $entityManager = $this->getDoctrine()->getManager();
                 $entityManager->remove($wallet);
@@ -163,6 +259,6 @@ class WalletController extends AbstractController
 
             return $this->redirectToRoute('wallet_index');
         }else
-            return $this->render('user_login');
+            return $this->redirectToRoute('user_login');
     }
 }
